refactor(tests): move repo-baseline shared fakes to conftest (#980)
Centralise FakeRunner, _jarvis_responses, _dependabot_b64, and _FakeProc in
tests/conftest.py (the established shared-test-infra home) and drop the
sibling-test coupling `from test_repo_baseline_auditor import ...` in
test_repo_baseline_generate_snapshots.py — a top-level import that only worked
under pytest prepend mode and would break silently on rename.

Kept as plain importable helpers (not fixtures) so all ~30 call sites stay
`FakeRunner(...)` / `_jarvis_responses()` byte-identical. Refreshed the
pyproject prepend-mode comment whose stale example referenced the now-removed
auditor-test import.

Closes #980
